fix: use proper vocab size for logit_bias
tokenizer.vocab_size returns the size of the vocab excluding added tokens while len(tokenizer) returns the proper size of the vocab including the added tokens. As seen here:
https://github.com/huggingface/transformers/blob/79254c9b61ac2e1a6e0f698c38b1654159095e2e/src/transformers/tokenization_utils_fast.py#L234-L239
https://github.com/huggingface/transformers/blob/79254c9b61ac2e1a6e0f698c38b1654159095e2e/src/transformers/tokenization_utils_fast.py#L275-L279

This prevents errors when using added tokens as logit bias, such as trying to bias eos_token_id on for example meta-llama/Llama-3.1-8B-Instruct

I am unsure if there are any other places where this should be changed, but looking at the use cases of vocab_size in the rest of the repo, they mostly seem to be derived from the model config, which *should* be the correct size.